refactor(core): lifecycle hooks in one single file
