--- packages/core/src/lifecycle/Lifecycle.ts
@@ -1,5 +1,0 @@
-interface ILifecycle {
-    run(): void;
-}
-
-export default ILifecycle;
